Fix TODO/compile warning - refactor DocumentGroup
